requirements.txt
Flask
pymongo
Flask-PyMongo
python-dotenv